test/web-platform-tests/to-upstream/encoding/meta/meta-http-equiv-reverse.html

<!DOCTYPE html>
<html>
  <head>
      <!--<meta http-equiv="Content-Type" content="text/html;charset=ISO-8859-8">-->
      <meta content="text/html;charset=ISO-8859-5" http-equiv="Content-Type"/>
      <script src="/resources/testharness.js"></script>
      <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      "use strict";
      test(() => {
        assert_equals(document.characterSet, "ISO-8859-5");
      }, "document.characterSet should be determined by meta charset");
    </script>
  </body>
</html>
